Created new config_template action plug-in
This commit adds a new action plug-in named "config_template". This
action plug-ins was created to allow overrides to be passed to
to standard templates written in "ini", "yaml", or "json" which are often
used for system configuration.

This allows a project to carry a configuration template without having to
create specific variables for all configuration items that could potentially
be used. This keeps configuration templates clean and simple but
enhances their ability to be used in different ways which is great for
developers and deployers alike as no two deployment environments are
ever the same.

Functional Overview:
  The action plug in functions much like the regular template module
  allowing an Ansible user to write the template containing dynamic
  data through the use of jinja variables. The source template is
  rendered in transit and passed through a particular configuration
  option method should any override data be present. The override
  data is presented to the module through a dictionary (hash). Once
  the overrides have been applied the template is rendered once again
  to ensure all jinja variables are appropriately rendered which gives
  the user the ability to use variables in the override data as well as
  the base template if they so choose. Finally, the resultant is passed
  through the copy module where it is transfered to disk.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>