Updated the ssh plugin to retry simple connections
The plugin is subject to errors that can leave a file descriptor
open causing the application to hang or otherwise fail to write
data. This change simply adds a retry decorator to the communicate
and and exec_command methods which should help the case where ssh
commands fail for random intermittent issues.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>